test mipmap lod

+++ rasterizer/math/mathf.inl
@@ -252,18 +252,18 @@ float Mathf::Exp(float f)
 
 float Mathf::Log2(float f)
 {
-    static float _inv_log2 = 1.f / ::log(2);
-    return ::log(f) * _inv_log2;
+    static float _inv_log2 = 1.f / ::logf(2.f);
+    return ::logf(f) * _inv_log2;
 }
     
 float Mathf::Log(float f, float d)
 {
-    return ::log(f) / ::log(d);
+    return ::logf(f) / ::logf(d);
 }
 
 float Mathf::Log10(float f)
 {
-	return ::log10(f);
+	return ::log10f(f);
 }
 
 bool Mathf::IsPowerOfTwo(int value)
